chore(FDS-438): [Boundaries] - move Stat and Title into atoms folder
