+++ packages/cascara/src/atoms/Stat/Stat.fixture.js
