+++ packages/cascara/src/atoms/index.js
@@ -7,3 +7,4 @@ export { default as Button } from './Button';
 export { default as Columns } from './Columns';
 export { default as Flex } from './Flex';
 export { default as Image } from './Image';
+export { default as Title } from './Title';
